requirements_pip.txt
--index-url https://download.pytorch.org/whl/cu118
--extra-index-url https://pypi.org/simple

torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118

transformers==4.57.6
datasets==4.5.0
accelerate==1.12.0

pandas==2.3.3
numpy==2.3.5
scikit-learn==1.8.0

psycopg2-binary==2.9.11
python-dotenv==1.2.1

